Move a kube-proxy unit test to the correct file
